+++ src/lib/utils/recruitment/dropOffCandidates.js
@@ -114,7 +114,7 @@ export function formatDropOffDate(value) {
   }
 
   return date.toLocaleDateString("en-PH", {
-    month: "short",
+    month: "long",
     day: "numeric",
     year: "numeric",
   });
