+++ src/lib/utils/weeklyReports/weeklyReportsHelpers.js
@@ -24,10 +24,10 @@ export function getCurrentWeekDateRange() {
   sunday.setDate(monday.getDate() + 6);
 
   return `${monday.toLocaleDateString("en-PH", {
-    month: "short",
+    month: "long",
     day: "numeric",
   })} - ${sunday.toLocaleDateString("en-PH", {
-    month: "short",
+    month: "long",
     day: "numeric",
     year: "numeric",
   })}`;
@@ -50,7 +50,7 @@ export function formatDate(date) {
   if (Number.isNaN(parsed.getTime())) return "—";
 
   return parsed.toLocaleDateString("en-PH", {
-    month: "short",
+    month: "long",
     day: "numeric",
     year: "numeric",
   });
